requirements.txt
playwright==1.48.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.31.0
aiohttp==3.9.5
